+++ tajma/forms/ConfirmForm.py
@@ -1,0 +1,5 @@
+from wtforms import SubmitField
+from flask_wtf import FlaskForm
+
+class ConfirmForm (FlaskForm):
+    confirm = SubmitField('Yes')
